Scrum 119 post processing ramió

+++ Survival/Assets/Scripts/Sistemes/SistemaVida/SistemaVidaEnemic.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class SistemaVidaEnemic : SistemaVida
 {
-    [SerializeField] private int vidaMaxima = 2;
-    [SerializeField] private int vidaActual = 2;
+    [SerializeField] private float vidaMaxima = 2f;
+    [SerializeField] private float vidaActual = 2f;
     
     // Referencias
     private Animator animator;
@@ -15,8 +15,8 @@ public class SistemaVidaEnemic : SistemaVida
     public event Action OnIniciarAtac;
     public event Action QuanMoriEnemic;
     // Acceso a propiedades
-    public int VidaActual => vidaActual;
-    public int VidaMaxima => vidaMaxima;
+    public float VidaActual => vidaActual;
+    public float VidaMaxima => vidaMaxima;
     
     public override void Awake()
     {
@@ -32,29 +32,24 @@ public class SistemaVidaEnemic : SistemaVida
         return vidaActual > 0;
     }
     
-    public void DecrementarVida(int quantitat, string font = "")
+    public void DecrementarVida(float quantitat, string font = "")
     {
-        // Evitamos modificar la vida si ya está muriendo
         if (quantitat <= 0 || !EsViu() || animator.GetBool("senseVida")) 
             return;
         
-        vidaActual = Mathf.Max(vidaActual - quantitat, 0);
+        vidaActual = Mathf.Max(vidaActual - quantitat, 0f);
         
-        // Registrar fuente de daño
         if (!string.IsNullOrEmpty(font))
             Debug.Log($"Enemic {name} rep {quantitat} de dany de {font}. Vida restant: {vidaActual}");
         
-        // Activamos la animación de recibir daño
         if (vidaActual > 0 && animator != null)
             animator.SetTrigger("RepMal");
         
-        // Notificamos el cambio de vida
         NotificarCanviVida();
         
-        // Si la vida llega a 0, iniciamos la muerte
         if (vidaActual <= 0)
         {
-            vidaActual = 0;
+            vidaActual = 0f;
             StartCoroutine(Morir());
         }
     }
